django-api/requirements.txt
django==1.10.3
django-cors-headers
djangorestframework==3.5.3
gunicorn==19.6.0
redis==2.10.5
